fix: Correct total_count type in SessionLogsResponse
- Updated the total_count field in the SessionLogsResponse struct to be explicitly cast to i64, ensuring type consistency and preventing potential overflow issues during analytics processing.

+++ crates/temps-analytics/src/analytics.rs
@@ -1089,7 +1089,7 @@ impl Analytics for AnalyticsService {
         Ok(Some(SessionLogsResponse {
             session_id,
             logs,
-            total_count: total_count,
+            total_count: total_count as i64,
             offset: offset_val as i32,
             limit: limit_val as i32,
         }))
